Mini timer App with swift
For beginner programmers You can pause the timer or reset the timer and best of all the time is remembered when the app is closed and reopened.

We are going to use user defaults in order to store the start and stop times. Using these dates we can calculate the time based on the difference in milliseconds between the current time and the saved times.